Refactored server-side organization for clarity
